pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wikiinsight-engine"
version = "0.1.0"
description = "AI-driven system for detecting and visualizing Wikipedia geographic bias gaps"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "WikiInsight Team"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]
dependencies = [
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "pyarrow>=12.0.0",
    "mwclient>=0.10.0",
    "requests>=2.31.0",
    "python-dateutil>=2.8.0",
    "sentence-transformers>=2.2.0",
    "spacy>=3.6.0",
    "networkx>=3.1",
    "igraph>=0.10.0",
    "pyvis>=0.3.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "lightgbm>=4.0.0",
    "imbalanced-learn>=0.11.0",
    "shap>=0.42.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.23.0",
    "pydantic>=2.0.0",
    "slowapi>=0.1.9",
    "plotly>=5.15.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "sqlalchemy>=2.0.0",
    "dvc[s3]>=3.0.0",
    "mlflow>=2.6.0",
    "prefect>=2.10.0",
    "great-expectations>=0.18.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "python-dotenv>=1.0.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "fairlearn>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.4.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

